+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Add duty ratio to `Modulation::Square`
 * Add `Controller::output_balance`
 * Del `wait_for_msg_processed` option to `Controller::send()`
-    * Add `Controller::hand_shake` instead
+    * Add `Controller::check_ack` instead
 * Update firmware version to 1.8
 
 # 1.7.2